chore: move rendering fn to own module
